+++ libs/elodin-editor/src/ui/command_palette/palette_items.rs
@@ -1,5 +1,5 @@
 use std::{
-    collections::{BTreeMap, HashMap},
+    collections::BTreeMap,
     path::{Path, PathBuf},
     str::FromStr,
     time::Duration,
@@ -298,7 +298,7 @@ pub fn create_graph(tile_id: Option<TileId>) -> PaletteItem {
 }
 
 fn graph_parts(
-    parts: &HashMap<String, eql::ComponentPart>,
+    parts: &BTreeMap<String, eql::ComponentPart>,
     tile_id: Option<TileId>,
 ) -> Vec<PaletteItem> {
     parts
@@ -361,7 +361,7 @@ pub fn create_monitor(tile_id: Option<TileId>) -> PaletteItem {
 }
 
 fn monitor_parts(
-    parts: &HashMap<String, eql::ComponentPart>,
+    parts: &BTreeMap<String, eql::ComponentPart>,
     tile_id: Option<TileId>,
 ) -> Vec<PaletteItem> {
     parts
@@ -605,6 +605,14 @@ fn set_time_range_behavior() -> PaletteItem {
     })
 }
 
+pub fn save_schematic_as() -> PaletteItem {
+    PaletteItem::new(
+        "Save Schematic As...",
+        PRESETS_LABEL,
+        |_name: In<String>| PalettePage::new(vec![save_schematic_inner()]).into(),
+    )
+}
+
 pub fn save_schematic() -> PaletteItem {
     PaletteItem::new(
         "Save Schematic",
@@ -621,7 +629,7 @@ pub fn save_schematic() -> PaletteItem {
                     }
                     PaletteEvent::Exit
                 }
-                None => PalettePage::new(vec![save_preset_inner()]).into(),
+                None => PalettePage::new(vec![save_schematic_inner()]).into(),
             }
         },
     )
@@ -629,7 +637,7 @@ pub fn save_schematic() -> PaletteItem {
 
 pub fn save_schematic_db() -> PaletteItem {
     PaletteItem::new(
-        "Save Schematic To Db",
+        "Save Schematic To DB",
         PRESETS_LABEL,
         |_name: In<String>, tx: Res<PacketTx>, schematic: Res<CurrentSchematic>| {
             let kdl = schematic.0.to_kdl();
@@ -644,7 +652,7 @@ pub fn save_schematic_db() -> PaletteItem {
     )
 }
 
-pub fn save_preset_inner() -> PaletteItem {
+pub fn save_schematic_inner() -> PaletteItem {
     PaletteItem::new(
         LabelSource::placeholder("Enter a name for the schematic"),
         "",
@@ -1073,8 +1081,8 @@ impl Default for PalettePage {
             create_sidebars(),
             create_3d_object(),
             save_schematic(),
-            // TODO: Don't enable until we have a clear save and load story for db.
-            //save_schematic_db(),
+            save_schematic_as(),
+            save_schematic_db(),
             load_schematic(),
             set_color_scheme(),
             PaletteItem::new("Documentation", HELP_LABEL, |_: In<String>| {