+++ libs/elodin-editor/src/ui/hierarchy.rs
@@ -1,4 +1,5 @@
 use crate::EqlContext;
+use crate::ui::{EntityFilter, EntityPair, SelectedObject, colors::get_scheme};
 use bevy::ecs::{
     system::{ResMut, SystemParam, SystemState},
     world::World,
@@ -7,9 +8,7 @@ use bevy::prelude::Res;
 use bevy_egui::egui;
 use fuzzy_matcher::{FuzzyMatcher, skim::SkimMatcherV2};
 use impeller2_bevy::EntityMap;
-use std::collections::HashMap;
-
-use crate::ui::{EntityFilter, EntityPair, SelectedObject, colors::get_scheme};
+use std::collections::BTreeMap;
 
 use super::{inspector::search, schematic::Branch, widgets::WidgetSystem};
 
@@ -166,7 +165,7 @@ fn component_part(
 }
 
 fn filter_component_parts<'a, 'b>(
-    children: &'b HashMap<String, eql::ComponentPart>,
+    children: &'b BTreeMap<String, eql::ComponentPart>,
     matcher: &SkimMatcherV2,
     str: &'a str,
 ) -> (Vec<(i64, String, &'b eql::ComponentPart)>, &'a str) {