+++ libs/impeller2/wkt/src/gui.rs
@@ -148,6 +148,8 @@ pub struct Split<T = ()> {
     pub panels: Vec<Panel<T>>,
     pub shares: HashMap<usize, f32>,
     pub active: bool,
+    #[serde(default)]
+    pub name: Option<String>,
 }
 
 impl<T> Split<T> {
@@ -156,6 +158,7 @@ impl<T> Split<T> {
             panels: self.panels.iter().map(|p| p.map_aux(&f)).collect(),
             shares: self.shares.clone(),
             active: self.active,
+            name: self.name.clone(),
         }
     }
 }
@@ -219,6 +222,7 @@ pub struct Graph<T = ()> {
     pub auto_y_range: bool,
     pub y_range: Range<f64>,
     pub aux: T,
+    pub colors: Vec<crate::Color>,
 }
 
 impl<T> Graph<T> {
@@ -230,6 +234,7 @@ impl<T> Graph<T> {
             auto_y_range: self.auto_y_range,
             y_range: self.y_range.clone(),
             aux: f(&self.aux),
+            colors: self.colors.clone(),
         }
     }
 }