+++ libs/nox/src/noxpr/mod.rs
@@ -1,3 +1,4 @@
+#![doc = include_str!("README.md")]
 pub mod batch_dfs;
 // We only use the batch module for testing purposes.
 #[cfg(test)]